Simplify config screen
