+++ install.php
@@ -65,14 +65,12 @@
                     'check',
                     'check/fail',
                     'config',
+                    'config/config',
                     'config/mysql',
                     'config/pgsql',
                     'config/sqlite',
                     'config/sqlsrv',
                     'config/fail',
-                    'config/database',
-                    'config/admin',
-                    'config/advanced',
                     'project',
                     'project/project',
                     'project/plugins',
