+++ install_files/partials/config.htm
@@ -2,24 +2,9 @@
 <div id="configForm" class="animated-content">
     <form id="configFormElement">
 
-        <div class="row">
-            <div class="col-md-3 hidden-sm">
-
-                <!-- Side Navigation -->
-                <div class="side-nav section-side-nav"></div>
-
-            </div>
-            <div class="col-md-9">
-
-                <div class="cloud-content">
-                    <!-- Configuration Form -->
-                    <div class="config-form-options form-horizontal section-content"></div>
-                </div>
-
-                <!-- Page Navigation -->
-                <div class="btn-group page-nav section-page-nav"></div>
-
-            </div>
+        <div class="cloud-content">
+            <!-- Configuration Form -->
+            <div class="config-form-options form-horizontal section-content"></div>
         </div>
 
     </form>
