--- install_files/partials/config/admin.htm
@@ -1,75 +1,0 @@
-<p class="lead">Please specify details for logging in to the Administration Area.</p>
-
-<div class="form-group">
-    <label for="adminFirstName" class="control-label col-sm-4">First Name</label>
-    <div class="col-sm-8">
-        <input
-            id="adminFirstName"
-            name="admin_first_name"
-            class="form-control"
-            value="Admin"
-            />
-    </div>
-</div>
-
-<div class="form-group">
-    <label for="adminLastName" class="control-label col-sm-4">Last Name</label>
-    <div class="col-sm-8">
-        <input
-            id="adminLastName"
-            name="admin_last_name"
-            class="form-control"
-            value="Person"
-            />
-    </div>
-</div>
-
-<div class="form-group">
-    <label for="adminEmail" class="control-label col-sm-4">Email Address</label>
-    <div class="col-sm-8">
-        <input
-            id="adminEmail"
-            name="admin_email"
-            class="form-control"
-            value="admin@domain.tld"
-            />
-    </div>
-</div>
-
-<div class="form-group">
-    <label for="adminLogin" class="control-label col-sm-4">Admin Login</label>
-    <div class="col-sm-8">
-        <input
-            id="adminLogin"
-            name="admin_login"
-            class="form-control"
-            value="admin"
-            />
-    </div>
-</div>
-
-<div class="form-group">
-    <label for="adminPassword" class="control-label col-sm-4">Admin Password</label>
-    <div class="col-sm-8">
-        <input
-            id="adminPassword"
-            type="password"
-            name="admin_password"
-            class="form-control"
-            value=""
-            />
-    </div>
-</div>
-
-<div class="form-group">
-    <label for="adminConfirmPassword" class="control-label col-sm-4">Confirm Password</label>
-    <div class="col-sm-8">
-        <input
-            id="adminConfirmPassword"
-            type="password"
-            name="admin_confirm_password"
-            class="form-control"
-            value=""
-            />
-    </div>
-</div>
