--- install_files/partials/config/advanced.htm
@@ -1,64 +1,0 @@
-<p class="lead">
-    Provide a custom URL for the Administration Area.
-</p>
-
-<div class="form-group">
-    <label for="advBackendUri" class="control-label col-sm-4">Backend URL</label>
-    <div class="col-sm-8">
-        <input
-            id="advBackendUri"
-            name="backend_uri"
-            class="form-control"
-            value="/backend" />
-        <span class="help-block">Please specify a value which you will use to access the Backend.</span>
-    </div>
-</div>
-
-<p class="lead">
-    Specify a unique code for protecting sensitive data, such as user passwords.
-</p>
-<div class="form-group">
-    <label for="advEncryptionCode" class="control-label col-sm-4">Encryption Code</label>
-    <div class="col-sm-8">
-        <input
-            id="advEncryptionCode"
-            name="encryption_code"
-            class="form-control"
-            value="RANDOM_STRING"
-            maxlength="32" />
-        <span class="help-block">The encryption code should be of a valid length (32).</span>
-    </div>
-</div>
-
-<p class="lead">
-    Specify a permission mask for folders and files used for installation and software updates.
-</p>
-
-<div class="form-group">
-    <label for="advFileMask" class="control-label col-sm-4">File Permission Mask</label>
-    <div class="col-sm-8">
-        <input
-            id="advFileMask"
-            name="file_mask"
-            class="form-control"
-            value="644" />
-    </div>
-</div>
-
-<div class="form-group">
-    <label for="advFolderMask" class="control-label col-sm-4">Folder Permission Mask</label>
-    <div class="col-sm-8">
-        <input
-            id="advFolderMask"
-            name="folder_mask"
-            class="form-control"
-            value="755" />
-    </div>
-</div>
-
-<p class="help-block">
-    * If any permission issues occur, then you may use <strong>777</strong> for folder permissions
-    and <strong>666</strong> for file permissions. Setting the folder or file permissions to <strong>
-    777</strong> will mean that anyone can write and/or execute anything in those folders or files. This
-    is <strong>not recommended</strong> for shared hosting.
-</p>
