+++ install_files/partials/config/config.htm
@@ -1,4 +1,22 @@
-<p class="lead">Please prepare an empty database for this installation.</p>
+<p class="lead">
+    Provide a custom URL for the Administration Area.
+</p>
+
+<div class="form-group">
+    <label for="advBackendUri" class="control-label col-sm-4">Backend URL</label>
+    <div class="col-sm-8">
+        <input
+            id="advBackendUri"
+            name="backend_uri"
+            class="form-control"
+            value="/backend" />
+        <span class="help-block">Please specify a value which you will use to access the Backend.</span>
+    </div>
+</div>
+
+<p class="lead">
+    Please prepare an empty database for this installation.
+</p>
 
 <div class="form-group">
     <label for="dbType" class="control-label col-sm-4">Database Type</label>
