+++ install_files/partials/config/fail.htm
@@ -1,4 +1,4 @@
 <div class="callout callout-danger">
     <h4>{{reason}}</h4>
-    <p>There is a problem with the specified {{label}} configuration.</p>
+    <p>There is a problem with the specified configuration.</p>
 </div>
