+++ install_files/partials/project.htm
@@ -9,9 +9,6 @@
                 <div class="section-content"></div>
             </div>
 
-            <!-- Page Navigation -->
-            <div class="btn-group page-nav section-page-nav"></div>
-
         </div>
         <div class="col-md-3 hidden-sm">
